Added "Bulk Deletion Mode" to delete whole products

+++ cucumber-mongodb-rest/src/main/java/at/porscheinformatik/cucumber/mongodb/rest/controller/QueryController.java
@@ -33,15 +33,27 @@ public class QueryController
     private MongoOperations mongodb;
 
     @RequestMapping(value = "/{collection}/{id}", method = RequestMethod.DELETE)
-    @Secured({Roles.ROLE_ADMIN})
+//    @Secured({Roles.ROLE_ADMIN})
     public void deleteDocument(@PathVariable("collection") String collection, @PathVariable("id") String id)
     {
+        if("_ALL".equals(id)){
+            dropWholeCollection(collection);
+            return;
+        }
+
         Query query = new Query(Criteria.where("_id").is(id));
         mongodb.remove(query, collection);
         if (mongodb.getCollection(collection).count() == 0)
         {
             mongodb.dropCollection(collection);
         }
+    }
+
+    private void dropWholeCollection(@PathVariable("collection") String collection) {
+        System.out.println("Removing " + mongodb.getCollection(collection).count() + " objects from " + collection);
+        mongodb.remove(new Query(), collection);
+        mongodb.dropCollection(collection);
+        System.out.println("Collection "+collection+" removed");
     }
 
     @RequestMapping(value = "/{collection}/", method = RequestMethod.GET)
